# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ConvocaFinder API
  version: 1.0.0
  description: API REST para gestionar convocatorias normalizadas y alertas.
servers:
  - url: http://localhost:8000
paths:
...
  /convocatorias:
    get:
      summary: Listar convocatorias
      parameters:
        - in: query
          name: country
          schema:
            type: string
        - in: query
          name: funding_type
          schema:
            type: string
        - in: query
          name: amount_min
          schema:
            type: number
        - in: query
          name: amount_max
          schema:
            type: number
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: size
          schema:
            type: integer
      responses:
        '200':
          description: Lista paginada de convocatorias
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedCalls'
  /convocatorias/{id}:
    get:
      summary: Obtener convocatoria
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Convocatoria
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CallRead'
        '404':
          description: No encontrado
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CallRead:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        organizer:
          type: string
        country:
          type: string
        deadline:
          type: string
          format: date
        funding_type:
          type: string
    PaginatedCalls:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/CallRead'
        pagination:
          type: object
          properties:
            total:
              type: integer
            page:
              type: integer
            size:
              type: integer
...